Add support for replaceable HTTP backends
Instead of tightly coupling to requests, allow the HTTP backend to be replaced with a custom implementation. Provide an abstract class and a default implementation that uses requests the same as it's used now.

Unit tests are included and have been run successfully on Python 2.7 and 3.10 locally.

The new functionality is documented in `README.md`, and mypy type annotations are included on new code.

**Compatibility note:** This change retains backward compatibility to Python 2.7 and deprecates Python 2.6 compatibility. Python 2.6 reached EOL on August 24, 2010, nearly 12 years ago, so I believe this change is safe to make.

**Legal:** This contribution was made in the course of my employment with Dropbox, Inc., which has authorized the release of this code under the MIT License. This code is offered without any warranty; see the terms of the MIT License in the `LICENSE` file for details. This license release is valid only if this commit passes PGP signature validation using the public key on the author's GitHub account.

Signed-Off-By: Dan Fuhry <fuhry@dropbox.com>